Avoid running oneskyDownload script during tests
- Move filterReadyTranslations to a new file
