+++ bin/oneskyDownload.js
@@ -3,26 +3,9 @@ import fs from 'fs';
 import path from 'path';
 import dotenv from 'dotenv';
 import oneSky from '@brainly/onesky-utils';
+import { filterReadyTranslations } from './translationHelpers';
 
 dotenv.config({ path: '.env.local' });
-
-export function filterReadyTranslations(translationsRaw, languagesRaw) {
-  const readyLanguages = JSON.parse(languagesRaw)
-    .data.filter(language => language.is_ready_to_publish)
-    .map(language => language.code);
-
-  const translations = JSON.parse(translationsRaw);
-
-  const content = readyLanguages.reduce(
-    (acc, language) => ({
-      ...acc,
-      [language]: translations[language],
-    }),
-    {},
-  );
-
-  return JSON.stringify(content);
-}
 
 async function downloadTranslations() {
   const options = {
